moved sensor- and cardmodel related stuff to CardFragment
